ggragged: Ragged Grids for 'ggplot2'

Extend 'ggplot2' facets to panel layouts arranged in a grid with ragged edges. facet_ragged_rows() groups panels into rows that can vary in length, facet_ragged_cols() does the same but for columns. These can be useful, for example, to represent nested or partially crossed relationships between faceting variables.

See also

Author

Maintainer: Mikko Marttila mikkmart@protonmail.com [copyright holder]